requirements.txt
pyserial
Flask
s